[WIP] fix dependencies
